MinimumQuantileInformation {mc2d}	R Documentation

Minimum Quantile Information Distribution

Description

Density, distribution function, quantile function and random generation for Minimum Quantile Information distribution.

Usage

dmqi(x, 
  mqi, 
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL, 
  k = 0.1, 
  intrinsic = NA,
  log = FALSE)

pmqi(q, 
  mqi, 
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL,
  k = 0.1,
  intrinsic = NA,
  lower.tail = TRUE,
  log.p = FALSE
)

qmqi(p, 
  mqi, 
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL, 
  k = 0.1, 
  intrinsic = NA,
  lower.tail = TRUE, 
  log.p = FALSE
)

rmqi(n, 
  mqi, 
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL, 
  k=0.1, 
  intrinsic = NA
)

pmqi(
  q,
  mqi,
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL,
  k = 0.1,
  intrinsic = NA,
  lower.tail = TRUE,
  log.p = FALSE
)

qmqi(
  p,
  mqi,
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL,
  k = 0.1,
  intrinsic = NA,
  lower.tail = TRUE,
  log.p = FALSE
)

rmqi(
  n,
  mqi,
  mqi.quantile = c(0.05, 0.5, 0.95),
  realization = NULL,
  k = 0.1,
  intrinsic = NA
)

Arguments

x, q
Vector of quantiles

mqi
Minimum quantile information

mqi.quantile
The quantile of ‘mqi'. It’s a vector of length 3. Default is 'c(0.05, 0.5, 0.95)', that is the 5th, 50th and 95th.

realization
Default is 'NULL'. If not 'NULL', used to define 'L' or 'U' (see details).

k
Overshot, default value is 0.1.

intrinsic
Use to specify a prior bounds of the intrinsic range. Default = 'NA'.

log, log.p
Logical; if 'TRUE', probabilities 'p' are given as 'log(p)'.

lower.tail
Logical; if 'TRUE' (default), probabilities are 'P[X <= x]' otherwise, 'P[X > x]'.

p
Vector of probabilities.

n
Number of observations.

Details

p_1, p_2, and p_3 are percentiles of a distribution with p_1 < p_2 < p_3. The interval [L,U] is given with:

L = x_p_1

U = x_p_3

The support of minimum quantile information distribution is determined by the intrinsic range:

[L^*, U^*] = [L - k*(U - L), U + k*(U - L)]

where k denotes an overshoot and is chosen by the analyst (usually k = 10%, which is the default value).

Given the three values of quantile, x_{p_1}, x_{p_2} and x_{p_3}, and define p_0 = 0, p_4 = 1, x_{p_0} = L^{*} and x_{p_4} = U^{*} the minimum quantile information distribution is given by:

Probability density function

f(x)=(p_i-p_(i-1))/(x_p_i-x_p_(i-1)) for x_p_(i-1)≤ x_p_i, i=1,…,4

f(x) = 0, otherwise

Cumulative distribution function

F(x) = 0 for x < x_p_0

F(x) = (p_i-p_(i-1))/(x_p_i-x_p_(i-1))*(x-x_p_(i-1))+p_(i-1) for x_p_(i-1)≤ x_p_i, i=1,…,4

F(x) = 1 for x_p_(4) ≤ x

This distribution is usually used for expert elicitation. If experts have realization information, then the range [L,U] is given by:

L = min(x_p_1, realization)

U = max(x_p_3, realization)

For some questions, experts may have information for the intrinsic range and set a prior intrinsic range (L^* and U^*).

NOTE that the function is vectorized only for x, q, p, n. As a consequence, it can't be used for variable other parameters.

Author(s)

Yu Chen and Arie Havelaar

References

Hanea, A. M., & Nane, G. F. (2021). An in-depth perspective on the classical model. In International Series in Operations Research & Management Science (pp. 225–256). Springer International Publishing.

Examples

curve(dmqi(x, mqi=c(40,50,60), intrinsic=c(0,100)), from=0, to=100, type = "l", xlab="x",ylab="pdf")
curve(pmqi(x, mqi=c(40,50,60), intrinsic=c(0,100)), from=0, to=100, type = "l", xlab="x",ylab="cdf")
rmqi(n = 10, mqi=c(555, 575, 586))

[Package mc2d version 0.2.0 Index]