dimmcnode {mc2d}	R Documentation

Dimension of mcnode and mc Objects

Description

Provides the dimension (i.e. the number of simulations in the variability dimension, the number of simulations in the uncertainty dimension and the maximum number of variates of a mcnode or a mc object.

Usage

dimmcnode(x)
dimmc(x)

Arguments

x
a mcnode or a mc object.

Value

A vector of three scalars: the dimension of variability (1 for "0" and "U" mcnode), the dimension of uncertainty (1 for "0" and "V" mcnode) and the number of variates (the maximal number of variates for an mc object.

Note

This function does not test if the object is correctly built. See is.mcnode and is.mc .

Examples

data(total)
dimmcnode(xVUM2)
dimmc(total)

[Package mc2d version 0.2.0 Index]