ec {mc2d}	R Documentation

An example on Escherichia coli in ground beef

Description

The fictive example is as following:

A batch of ground beef is contaminated with E. coli, with a mean concentration conc.

Consumers may eat the beef "rare", "medium rare" or "well cooked". If "rare", no bacteria is killed. If "medium rare", 1/5 of bacteria survive. If "well cooked", 1/50 of bacteria survive.

The serving size is variable.

The risk of infection follows an exponential model.

For the one-dimensional model, it is assumed that:

conc <- 10

cook <- sample(n, x=c(1,1/5,1/50),replace=TRUE,prob=c(0.027,0.373,0.600))

serving <- rgamma(n, shape=3.93,rate=0.0806)

expo <- conc * cook * serving

dose <- rpois(n, lambda=expo)

risk <- 1-(1-0.001)^dose

For the two-dimensional model, it is assumed moreover that the concentration and the r parameter of the dose response are uncertain.

conc <- rnorm(n,mean=10,sd=2)

r <- runif(n ,min=0.0005,max=0.0015)

Usage

data(ec)

Format

A list of two expression to be passed in mcmodel

Source

Fictive example

References

None

[Package mc2d version 0.2.0 Index]