evalmcmod {mc2d}	R Documentation

Evaluates a Monte-Carlo model

Description

Evaluates a mcmodel object (or a valid expression) using a specified number of simulations and with (or without) a specified seed.

Usage

evalmcmod(expr, nsv=ndvar(), nsu=ndunc(), seed=NULL)

Arguments

expr
A model of class mcmodel or a valid expression.

nsv
The number of simulations in the dimension of variability used in the evaluation.

nsu
The number of simulations in the dimension of uncertainty used in the evaluation.

seed
The random seed used for the evaluation. If NULL the seed is unchanged.

Details

The model is evaluated. The intermediate variables used to build the mc object are not stored.

Value

The results of the evaluation. It should be a mc object.

Note

The seed is set at the beginning of the evaluation. Thus, the complete similarity of two evaluations with similar seed is not certain, depending on the structure of your model.

See Also

mcmodel

evalmccut to evaluate high dimension Monte Carlo Model in a loop.

Examples

data(ec)
ec$modEC1
evalmcmod(ec$modEC1, nsv=100, nsu=100, seed=666)

[Package mc2d version 0.2.0 Index]