extractvar {mc2d}	R Documentation

Utilities for multivariate nodes

Description

extractvar extracts one variate from a multivariate node.

addvar adds consistent mcnodes to build a multivariate mcnode .

Usage

extractvar(x, which=1)
addvar(...)

Arguments

x
a multivariates mcnode.

which
a vector. which variate(s) should be extracted?

...
mcnodes to be gathered in a multivariate mcnode. These mcnodes should be of same type and dimension.

Details

The outm attribute of the output of addvar will be the one of the first element.

Value

The new built mcnode.

See Also

mcnode for mcnode objects.

Examples

x <- mcdata(0:3, "0", nvariates = 4)
y <- extractvar(x, c(1, 3)) 
y
addvar(x, y)

[Package mc2d version 0.2.0 Index]