ggtornado {mc2d}	R Documentation

Draws a Tornado chart as provided by tornado (ggplot version).

Description

Draws a Tornado chart as provided by tornado.

Usage

## For class 'tornado'
ggtornado(x, 
  which=1, 
  name=NULL, 
  stat=c("median","mean"), 
  xlab="method", 
  ylab=""
)

## For class 'tornadounc'
ggtornadounc(x,
  which=1, 
  stat="median", 
  name=NULL, 
  xlab="method", 
  ylab=""
)

ggtornadounc(
  x,
  which = 1,
  stat = "median",
  name = NULL,
  xlab = "method",
  ylab = ""
)

Arguments

x
A tornado object as provided by the tornado function.

which
Which output to print -for multivariates output-.

name
Vector of name of input variables. If NULL, the name will be given from the name of the elements.

stat
The name of the statistics of the output to be considered. For a tornado object: "median" or "mean". For a tornadounc object: the value should match one row name of the tornadounc object. Alternatively, for a tornadounc object, the number of the row may be used.

xlab
Label of the x axis. Default is to use the correlation method used in the tornado object.

ylab
Label of the y axis. Default is empty.

See Also

tornado

Examples

data(ec)
x <- evalmcmod(ec$modEC2, nsv=100, nsu=100, seed=666)
tor <- tornado(x, 7)
ggtornado(tor)
data(total)
ggtornado(tornadounc(total, 10, use="complete.obs"), which=1)

[Package mc2d version 0.2.0 Index]