hist.mc {mc2d}	R Documentation

Histogram of a Monte Carlo Simulation

Description

Shows histogram of a mcnode or a mc object.

Usage

## S3 method for class 'mc'
hist(x, griddim=NULL, xlab=names(x), ylab="Frequency", main="", ...)
## S3 method for class 'mcnode'
hist(x, ...)

Arguments

x
An mcnode or an mc object.

griddim
A vector of two integers, indicating the size of the grid of plots. If NULL, the grid is calculated to produce a "nice" graph.

xlab
A vector of labels for the x-axis for drawn histograms (those whose outm(x)!="none"). May be recycled.

ylab
A vector of labels for the y-axis for drawn histograms. May be recycled.

main
A vector of main title of histograms for drawn histograms. May be recycled.

...
Other arguments to be passed to all calls of hist.

Note

For Two-dimensional mc, the histogram is based on all data (variability and uncertainty) pooled together.

Examples

data(total)
hist(xVUM3)
hist(total)

[Package mc2d version 0.2.0 Index]