is.mc {mc2d}	R Documentation

Tests mc and mcnode Objects

Description

is.mc tests mc objects and is.mcnode tests mcnode objects.

Usage

is.mc(x)
is.mcnode(x)

Arguments

x
An mc or a mcnode object.

Details

is.mc tests if x is a list of mcnode, each elements being of compatible dimension. It tests if the class "mc" is affected to the object.

is.mcnode tests if x is an array of numeric or logical, if it has a "type" attribute and compatible dimensions, and if the class "mcnode" is affected to the object.

Value

TRUE or FALSE

Examples

data(total)
is.mcnode(xVU)
is.mcnode(total)
is.mc(total)

[Package mc2d version 0.2.0 Index]