mc {mc2d}	R Documentation

Monte Carlo Object

Description

Creates mc objects from mcnode or mc objects.

Usage

mc(..., name=NULL, devname=FALSE)

Arguments

...
mcnode and/or mc object(s) to be gathered in a mc object separated by a coma.

name
Vector of character of the same length of the final mc object. If NULL, the name will be given from the name of the elements.

devname
Develop the name from the name of the mc objects, if any.

Details

A mc object is a list of mcnode objects. mcnode objects must be of coherent dimensions.

If one of the arguments is a mc object, the name of the elements of this mc object are used. devname = TRUE will develop the name, using as a prefix the name of the mc object.

Finally, names are transformed to be unique.

Value

An object of class mc.

See Also

mcnode, the basic element of a mc object.

To evaluate mc objects: mcmodel, evalmcmod, evalmccut

Informations about an mc object: is.mc, dimmc

To study mc objects: print.mc, summary.mc, plot.mc, converg, hist.mc, tornado, tornadounc.mc

Examples

 
x <- mcstoc(runif)
y <- mcdata(3, type="0")
z <- x * y
(m <- mc(x, y, z, name=c('n1', 'n2', 'n3')))
mc(m, x, devname=TRUE)

[Package mc2d version 0.2.0 Index]