mcratio {mc2d}	R Documentation

Ratio of uncertainty and the variability

Description

Provides measures of variability, uncertainty, and both combined for an mc or an mcnode object.

Usage

mcratio(x, pcentral=.5, pvar=.975, punc=.975, na.rm=FALSE)

Arguments

x
an mc or an mcnode object

pcentral
the quantile for the central tendency.

.

pvar
the quantile for the measure of variability.

punc
the quantile for the measure of uncertainty.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Details

The function evaluates three ratios for each mcnode. Given:

A

the (100 * pcentral)th percentile of uncertainty (by default the median) for the (100 * pcentral)th percentile of variability

B

the (100 * pcentral)th percentile of uncertainty for the (100 * pvar)th percentile of variability

C

the (100 * punc)th percentile of uncertainty for the (100 * pcentral)th percentile of variability

D

the (100 * punc)th percentile of uncertainty for the (100 * pvar)th percentile of variability

The following ratio are estimated

Variability Ratio: B / A

Uncertainty Ratio: C / A

Overall Uncertainty Ratio: D / A

For multivariate nodes, the statistics are evaluate on each dimension or on statistics according to the corresponding outm value.

Value

A matrix.

References

Ozkaynak, H., Frey, H.C., Burke, J., Pinder, R.W. (2009) "Analysis of coupled model uncertainties in source-to-dose modeling of human exposures to ambient air pollution: A PM2.5 case study", Atmospheric environment, Volume 43, Issue 9, March 2009, Pages 1641-1649.

Examples

data(total)
mcratio(total, na.rm=TRUE)

[Package mc2d version 0.2.0 Index]