plot.tornado {mc2d}	R Documentation

Draws a Tornado chart.

Description

Draws a Tornado chart as provided by tornado.

Usage

## S3 method for class 'tornado'
plot(x, which=1, name=NULL, stat=c("median", "mean"), xlab="method",
	  ylab="", ...)
## S3 method for class 'tornadounc'
plot(x, which=1, stat="median", name=NULL, xlab="method", ylab="", ...)

Arguments

x
A tornado object or a tornadounc object.

which
Which output to print -for multivariates output-.

name
Vector of name of input variables. If NULL, the name will be given from the name of the elements.

stat
The name of the statistics of the output to be considered. For a tornado object: "median" or "mean". For a tornadounc object: the value should match one row name of the tornadounc object. Alternatively, for a tornadounc object, the number of the row may be used.

xlab
Label of the x axis. if "method", use the correlation method used in the tornado object.

ylab
Label of the y axis.

...
Further arguments to be passed to the plot function.

Details

A point is drawn at the estimate and the segment reflects the uncertainty around this estimate.

Value

NULL

See Also

tornado

Examples

data(ec)
x <- evalmcmod(ec$modEC2, nsv=100, nsu=100, seed=666)
tor <- tornado(x, 7)
plot(tor)

[Package mc2d version 0.2.0 Index]