print.mc {mc2d}	R Documentation

Prints a mcnode or a mc Object

Description

Print a description of the structure of the mc or the mcnode object.

Usage

## S3 method for class 'mc'
print(x, digits=3, ...)
## S3 method for class 'mcnode'
print(x, ...)

Arguments

x
a mcnode or a mc object.

digits
Number of digits to be used.

...
Further arguments to be passed to the print function.

Value

An invisible data frame.

See Also

mcnode for mcnode objects. mc for mc objects.

[Package mc2d version 0.2.0 Index]