spaghetti {mc2d}	R Documentation

Spaghetti Plot of mc/mcnode Object

Description

Use plot to draw spaghetti plots for the mc/mcnode objects.

Usage

spaghetti(x, ...)

## S3 method for class 'mc'
spaghetti(
  x,
  griddim = NULL,
  xlab = names(x),
  ylab = "F(n)",
  main = "",
  maxlines = 100,
  ...
)

## S3 method for class 'mcnode'
spaghetti(x, ...)

Arguments

x
mc/mcnode object

...
further arguments to be passed to plot.stepfun()

griddim
a vector of two integers, indicating the size of the grid of the graph. If NULL, the grid is calculated to produce a "nice" graph.

xlab
vector of labels for the x-axis. If NULL, use the name of the node.

ylab
vector of labels for the y-axis.

main
vector of main titles of the graph.

maxlines
the maximum number of ecdf to draw.

Examples

data(total)
spaghetti(mc(xVUM))
spaghetti(xVUM)

[Package mc2d version 0.2.0 Index]