summary.mc {mc2d}	R Documentation

Summary of mcnode and mc Object

Description

Provides a summary of a mcnode, a mc or a mccut object.

Usage

## S3 method for class 'mc'
summary(object, probs=c(0, 0.025, 0.25, 0.5, 0.75, 0.975, 1), lim=c(0.025,
	  0.975), ...)
## S3 method for class 'mcnode'
summary(object, probs=c(0, 0.025, 0.25, 0.5, 0.75, 0.975, 1), lim=c(0.025,
	  0.975), digits=3, ...)
## S3 method for class 'mc'
print.summary(x, digits=3, ...)
## S3 method for class 'mccut'
summary(object, lim=c(0.025, 0.975), ...)

Arguments

object
a mcnode or a mc object or a mccut object.

x
A summary.mc object as provided by the summary.mc function.

probs
A vector of values used for the quantile function (variability dimension).

digits
Number of digits in the print.

lim
A vector of values used for the quantile function (uncertainty dimension).

...
For generic functions consistency.

Details

The mean, the standard deviation and the probs quantiles will be evaluated in the variability dimension. The median, the mean and the lim quantiles will then be evaluated on these statistics in the uncertainty dimension.

Multivariate nodes:

If the "outm" attributes of the mcnode is "none", the node is not evaluated, if it is "each" the variates are evaluated one by one, if it is a function (e.g. "mean"), the function is applied on the nvariates dimension before providing a classical output.

Value

a list.

See Also

mcnode for mcnode objects, mc for mc objects, mccut for mccut objects, quantile

Examples

data(total)
summary(xVUM3)
summary(total)

[Package mc2d version 0.2.0 Index]