typemcnode {mc2d}	R Documentation

Provides the Type of a mcnode Object

Description

Provide the type of a mcnode object.

Usage

typemcnode(x, index=FALSE)

Arguments

x
a mcnode object

index
if TRUE give the index of the type rather than the type.

Value

"0", "V","U" or "VU" or the corresponding index if index=TRUE.

NULL if none of this element is found.

Note

This function does not test if the object is correct. See is.mcnode.

Examples

data(total)
typemcnode(total$xVUM2)

[Package mc2d version 0.2.0 Index]